+++ src/main/webapp/resources/js/components/ant.design/table-renderers.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Button, Icon } from "antd";
+import { Icon } from "antd";
 import { formatInternationalizedDateTime } from "../../utilities/date-utilities";
 import { blue6 } from "../../styles/colors";
 
@@ -7,15 +7,15 @@ export const idColumnFormat = () => ({
   dataIndex: "id",
   key: "identifier",
   sorter: true,
-  width: 50
+  width: 120
 });
 
-export const nameColumnFormat = ({ url, width = 300 }) => {
+export const nameColumnFormat = ({ url }) => {
   return {
     dataIndex: "name",
     key: "name",
     sorter: true,
-    width,
+    ellipsis: true,
     filterIcon(filtered) {
       return (
         <Icon
@@ -28,26 +28,9 @@ export const nameColumnFormat = ({ url, width = 300 }) => {
     },
     render(name, data) {
       return (
-        <Button
-          type="link"
-          className="t-name"
-          href={`${url}/${data.id}`}
-          title={name}
-          style={{
-            textAlign: "left"
-          }}
-        >
-          <span
-            style={{
-              whiteSpace: "nowrap",
-              overflow: "hidden",
-              textOverflow: "ellipsis",
-              width: width - 50
-            }}
-          >
-            {name}
-          </span>
-        </Button>
+        <a className="t-name" href={`${url}/${data.id}`} title={name}>
+          {name}
+        </a>
       );
     }
   };