+++ src/main/webapp/resources/js/pages/projects/linelist/components/TableControlPanel/TemplatesPanel/TemplateSelect/TemplateSelect.jsx
@@ -7,7 +7,6 @@ import {
   PopoverContents
 } from "../../../../../../../components/popovers/index";
 
-
 const { Option } = Select;
 
 /*